Fix last IMC element not found to scroll
It should scroll down to the last IMC that we are looking for with the appropriate CSS selector but sometimes the last IMC element does not found and throws error. So, just wrapped it with try-catch so that the program continues with the found IMCs. For example, if we are looking for 40 IMCs but found 30 then continue.