convert to loop for scrolling down to find IMCs.
It can scroll down as much as needed until it finds proper number of IMCs it needs. First, it scroll down to 30th chats and if it does not find the total number of IMCs then it will scroll down to 50th (increase 20 on each scroll) chat to find the total IMCs and so on.